restored T0 1.0.0 version

+++ lib/AFE-Data-Structures/AFE-Data-Structures.h
@@ -100,7 +100,7 @@ struct DS18B20 {
   uint8_t gpio;
   float correction;
   uint16_t interval;
-  uint8_t unit;
+  uint16_t unit;
 };
 
 struct DOMOTICZ {
